requirements.txt
# Qwen2.5 Chat Completion Service - Python Dependencies

# Web framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# HTTP client for tokenizer/model communication
httpx>=0.25.0
requests>=2.31.0

# Configuration
pyyaml>=6.0.1

# Data validation
pydantic>=2.5.0

# Async support
aiofiles>=23.2.1

# Optional: Better JSON performance
orjson>=3.9.10

# Development/testing (optional)
pytest>=7.4.3
pytest-asyncio>=0.21.1
black>=23.11.0
